api/services/comment/requirements.txt
uvicorn
fastapi
pydantic
psycopg2-binary
pyjwt
